Big Commit - Shield Enemy Logic Complete + Bug fixing
- Fixed more bugs related to damage/health system
- Added Direct + Splash damage to drone homing projectiles with debug visualisation
- Added Shield Enemy logic + Minigun hitscan logic + Slam shockwave logic + placeholder animations for shield enemy as well as death animations for all enemies

+++ Assets/Script/Guns/BulletPoolManager.cs
@@ -33,7 +33,7 @@ public class BulletPoolManager : MonoBehaviour
         Bullet.GetComponent<ProjectileBase>().onBulletHit += returnBulletToPool;
     }
 
-    public void ShootBullet(Vector3 direction, Vector3 origin, float speed, int dmg)
+    public void ShootBullet(Vector3 direction, Vector3 origin, float speed, float dmg)
     {
         if (bullets.Count <= 0)
         {
